Resolve issues with using lambdas for message texts
It was possible to use lambdas (or to be complete: a MessageResolver) to provide
 the message at execution time of the prompt, which allowed using the
conversation context to influence the text. Unfortunately, the compiler seems to
 have a hard time working with such a construct and often could not infer the
type resulting in errors. This resolves it by basing the resolver on a more
generic function type, which is also better for integrating given the now very
common function type.