Add functionality to jump to given dialog
This adds `continueWith()` functionality which allows one to continue with a
specific prompt or prompt chain. This is helpful when wanting to return to a
much earlier prompt in the chain (e.g. restart a whole sequence).

For this to work, we now return the created prompts upon configuring them so the
 user can save them in a variable. Later, these can then be passed to
`continueWith()` to set that prompt as the next.

    buildPrompts {
        val start = message("This is the start")
        continueWith(start)
    }

It's important to note that - as visible in the example above - this can create
endless loops. The `retry()` mechanic is still necessary (and in many cases
also more elegant) because of how scoping works. For example, the following will
 not compile:

    buildPrompts {
        val input = text("Mesage") { _, _ ->
            continueWith(input)
        }
    }

Because `input` is not in scope for the callback. In these cases `retry()` is
still necessary.

What `continueWith()` also allows us to do, is to connect separate dialogs as
shown below.

    val second = buildPrompts {
        message("Second dialog")
    }

    buildPrompts {
        message("First dialog")
        continueWith(second)
    }

Additionally, this improves readability for long dialog chains by allowing the
user to split a large dialog into logical chunks.